add filesystem monitoring
